Annotated data: Move preexisting tests to proper location
